requirements.txt
streamlit==1.24.0
pillow-heif==0.11.1
